+++ src/java/com/onlinegaragesale/repository/GenericDAO.java
@@ -1,6 +1,7 @@
 package com.onlinegaragesale.repository;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * File Name: GenericDAO.java
@@ -11,7 +12,7 @@ import java.io.Serializable;
  * Date: 08 Sep 2012
  * Edited: 08 Sep 2012
  */
-public interface GenericDAO<T extends Serializable> extends DAO<T, Long>
+public interface GenericDAO<T extends Serializable> extends DAO<T, BigDecimal>
 {
 
     void setClazz(final Class<T> clazzToSet);
