+++ src/java/com/onlinegaragesale/services/crud/ProductsalesCrudService.java
@@ -2,6 +2,7 @@ package com.onlinegaragesale.services.crud;
 
 import com.onlinegaragesale.model.Productsales;
 import com.onlinegaragesale.services.Service;
+import java.math.BigDecimal;
 
 /**
  * File Name: StudentCrudService.java
@@ -12,6 +13,6 @@ import com.onlinegaragesale.services.Service;
  * Date: 08 Sep 2012
  * Edited: 08 Sep 2012
  */
-public interface ProductsalesCrudService extends Service<Productsales, Long>
+public interface ProductsalesCrudService extends Service<Productsales, BigDecimal>
 {
 }
