+++ src/java/com/onlinegaragesale/services/crud/impl/ProductsalesCrudServiceImpl.java
@@ -3,6 +3,7 @@ package com.onlinegaragesale.services.crud.impl;
 import com.onlinegaragesale.model.Productsales;
 import com.onlinegaragesale.repository.GenericDAO;
 import com.onlinegaragesale.services.crud.ProductsalesCrudService;
+import java.math.BigDecimal;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -36,7 +37,7 @@ public class ProductsalesCrudServiceImpl implements ProductsalesCrudService
     }
 
     @Override
-    public Productsales findById(Long id)
+    public Productsales findById(BigDecimal id)
     {
         setDao(dao);
         return dao.findById(id);
@@ -71,7 +72,7 @@ public class ProductsalesCrudServiceImpl implements ProductsalesCrudService
     }
 
     @Override
-    public void removeById(Long entityId)
+    public void removeById(BigDecimal entityId)
     {
         setDao(dao);
         Productsales v = dao.findById(entityId);
