+++ src/java/com/onlinegaragesale/services/crud/impl/SaleshistoryCrudServiceImpl.java
@@ -3,6 +3,7 @@ package com.onlinegaragesale.services.crud.impl;
 import com.onlinegaragesale.model.Saleshistory;
 import com.onlinegaragesale.repository.GenericDAO;
 import com.onlinegaragesale.services.crud.SaleshistoryCrudService;
+import java.math.BigDecimal;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -36,7 +37,7 @@ public class SaleshistoryCrudServiceImpl implements SaleshistoryCrudService
     }
 
     @Override
-    public Saleshistory findById(Long id)
+    public Saleshistory findById(BigDecimal id)
     {
         setDao(dao);
         return dao.findById(id);
@@ -71,7 +72,7 @@ public class SaleshistoryCrudServiceImpl implements SaleshistoryCrudService
     }
 
     @Override
-    public void removeById(Long entityId)
+    public void removeById(BigDecimal entityId)
     {
         setDao(dao);
         Saleshistory v = dao.findById(entityId);
